refactor: Move mapping models to repo layer
Remove Twirl plugin dependency and migrate mapping-related classes
(MappingElement, MappingStandoffDatatypeClass, MappingXMLAttribute, and
SparqlTemplateLinkUpdate) from org.knora.webapi.messages.twirl to
org.knora.webapi.slice.resources.repo.model. Delete unused Contributor
class and properties.scala.html template. Update all imports across
responders, repositories, and tests. Remove TxtFormat helper methods from
TriplestoreService that were only used for Twirl templates.
