+++ build.sbt
@@ -187,7 +187,7 @@ lazy val webapi: Project = Project(id = "webapi", base = file("webapi"))
     Test / parallelExecution  := true, // run tests in parallel
     libraryDependencies ++= Dependencies.webapiDependencies ++ Dependencies.webapiTestDependencies,
   )
-  .enablePlugins(SbtTwirl, JavaAppPackaging, DockerPlugin, JavaAgent, BuildInfoPlugin, HeaderPlugin)
+  .enablePlugins(JavaAppPackaging, DockerPlugin, JavaAgent, BuildInfoPlugin, HeaderPlugin)
   .settings(
     name := "webapi",
     resolvers ++= Seq(
