+++ webapi/src/main/scala/org/knora/webapi/slice/resources/repo/ChangeLinkMetadataQuery.scala
@@ -16,11 +16,11 @@ import zio.IO
 import java.time.Instant
 
 import dsp.errors.SparqlGenerationException
-import org.knora.webapi.messages.twirl.SparqlTemplateLinkUpdate
 import org.knora.webapi.slice.api.admin.model.Project
 import org.knora.webapi.slice.common.KnoraIris.ResourceIri
 import org.knora.webapi.slice.common.QueryBuilderHelper
 import org.knora.webapi.slice.common.repo.rdf.Vocabulary.KnoraBase as KB
+import org.knora.webapi.slice.resources.repo.model.SparqlTemplateLinkUpdate
 import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Update
 
 /**
