+++ webapi/src/main/scala/org/knora/webapi/slice/resources/repo/CreateLinkQuery.scala
@@ -23,11 +23,11 @@ import java.util.UUID
 
 import dsp.errors.SparqlGenerationException
 import dsp.valueobjects.UuidUtil
-import org.knora.webapi.messages.twirl.SparqlTemplateLinkUpdate
 import org.knora.webapi.slice.api.admin.model.Project
 import org.knora.webapi.slice.common.KnoraIris.ResourceIri
 import org.knora.webapi.slice.common.QueryBuilderHelper
 import org.knora.webapi.slice.common.repo.rdf.Vocabulary.KnoraBase as KB
+import org.knora.webapi.slice.resources.repo.model.SparqlTemplateLinkUpdate
 
 /**
  * Inserts a new link between resources.
