--- webapi/src/main/twirl/org/knora/webapi/messages/twirl/views/resource/properties.scala.html
@@ -1,25 +1,0 @@
-@*
-* Copyright © 2021 - 2024 Swiss National Data and Service Center for the Humanities and/or DaSCH Service Platform contributors.
-* SPDX-License-Identifier: Apache-2.0
-*@
-
-@(properties: Map[String, String], imgpath: String)
-
-<table>
-
-    @if(imgpath.nonEmpty) {
-        <tr>
-            <td style="border: none">preview</td>
-            <td style="border: none">:</td>
-            <td style="border: none"><img src="@imgpath" style="max-width: 128px; max-height: 128px;"></td>
-        </tr>
-    }
-
-    @for((k, v) <- properties) {
-    <tr>
-        <td style="border: none">@k</td>
-        <td style="border: none">:</td>
-        <td style="border: none">@v</td>
-    </tr>
-    }
-</table>
